move percentile value mapping out of utils
